Don't delete on forward
* when we have an incoming delete toward an owner, don't delete at
  that point; delete on the flood back.

* Test to make sure that the SyncDeleted record is created in a forwarded delete

+++ entanglement/sql/base.py
@@ -328,7 +328,8 @@ class SqlSyncRegistry(interface.SyncRegistry):
         ins = inspect(obj)
         if not ins.persistent: return
         assert obj in session
-        session.delete(obj)
+        if obj.sync_is_local or obj.sync_owner.destination == sender:
+            session.delete(obj)
         if obj.sync_owner and obj.sync_owner.destination == sender:
             #It's from the sender.
             sd = SyncDeleted(sync_serial = obj.sync_serial,
